small refactorings

+++ MotorSetupController.py
@@ -14,19 +14,24 @@ class MotorSetupController:
         # assert issubclass(cls, Steuerung)
         if not ACTIVE_SETTINGS_PATH.is_file():
             if STANDARD_SETTINGS_PATH.is_file():
-                ACTIVE_SETTINGS_PATH = STANDARD_SETTINGS_PATH
+                os.symlink(STANDARD_SETTINGS_PATH, ACTIVE_SETTINGS_PATH)
+                # ACTIVE_SETTINGS_PATH = STANDARD_SETTINGS_PATH
             else:
                 raise Exception(f"Active and standard settings paths are broken: {ACTIVE_SETTINGS_PATH}, {STANDARD_SETTINGS_PATH}")
-        self.currentDict = json.loads(open(ACTIVE_SETTINGS_PATH, encoding="utf8").read())
+        settingsString = open(ACTIVE_SETTINGS_PATH, encoding="utf8").read()
+        self.newDict = json.loads(settingsString)
+        self.currentDict = json.loads(settingsString)
             
-        self.current = cls(initializeMotors=self.currentDict)
+        self.current = cls(initializeMotors=self.newDict)
         self.cls = cls
         self.log = log
 
-    def reloadWithNewConfig(self, force: bool = False):        
+    def reloadWithNewConfig(self, config = None, force: bool = False):        
         self.current.stop()
+        if config is not None:
+            self.updateDict(config)
         try:
-            newSteuerung = self.cls(initializeMotors=self.currentDict)
+            newSteuerung = self.cls(initializeMotors=self.newDict)
         except Exception as e:
             self.log(f"New Steuerung could not be setup properly with the transmitted motor config. You can force it. Exception: {e}")
             # try:
@@ -35,7 +40,8 @@ class MotorSetupController:
             #     self.log("ERROR ALSO WHEN RESETTING")
             # return
             if not force: return False
-
+        self.currentDict = self.newDict
+        self.current = newSteuerung
         try:
             self.current.schreiben("CLOSING DUE TO UPDATED MOTOR CONFIG")
             self.current.quit()
@@ -43,40 +49,42 @@ class MotorSetupController:
             self.log("Old Steuerung could not be stopped properly.  Exception: " + e)# You can force it.
             # if not force:
             #     return
-        self.current = newSteuerung
         return True
+
+    def updateConfig(self, path: Path = ACTIVE_SETTINGS_PATH):
+        self.newDict = json.loads(open(path), encoding="utf8").read()
 
         
     def load(self, newDict: dict, force):
-        self._updateDict(newDict) # TODO Types are missing
         if not self.reloadWithNewConfig(force=force):
             return
         self.current.schreiben("RESTARTED WITH NEW MOTOR CONFIG. ")
         self.current.schreiben("New config: " + str(newDict), 3)
         newPath = CONFIGS_PATH / datetime.datetime.now().strftime('Config %Y.%m.%d_%H_%M_%S.json')
-        json.dump(self.currentDict, newPath.open('w', encoding='utf8'))
+        json.dump(self.newDict, newPath.open('w', encoding='utf8'))
         os.remove(ACTIVE_SETTINGS_PATH)
         os.symlink(newPath, ACTIVE_SETTINGS_PATH)
-        self.log("Successfully set up new InternetSteuerung with the transmitted motor config!")
+        self.log(f"Successfully set up new {self.cls.__name__} with the transmitted motor config!")
 
     def reset(self, force: bool = False):
-        self.currentDict = json.loads(open(STANDARD_SETTINGS_PATH, encoding="utf8").read())
-        self.reloadWithNewConfig()
-        self.current.schreiben("RESTARTED WITH DEFAULT MOTOR CONFIG. ")
-        self.current.schreiben("Default config: " + str(self.currentDict), 3)
-        os.remove(ACTIVE_SETTINGS_PATH)
-        os.symlink(STANDARD_SETTINGS_PATH, ACTIVE_SETTINGS_PATH)
+        self.updateConfig(STANDARD_SETTINGS_PATH)
+        if self.reloadWithNewConfig(force=force):
+            self.current.schreiben("RESTARTED WITH DEFAULT MOTOR CONFIG. ")
+            self.current.schreiben("Default config: " + str(self.newDict), 3)
+            os.remove(ACTIVE_SETTINGS_PATH)
+            os.symlink(STANDARD_SETTINGS_PATH, ACTIVE_SETTINGS_PATH)
 
-    def _updateDict(self, newDict):
+    def updateDict(self, newDict):
+        # TODO Types are missing
         for key in newDict:
             c = newDict[key]
             assert isinstance(c, dict)
-            if not key in self.currentDict:
-                self.currentDict[key] = {}
+            if not key in self.newDict:
+                self.newDict[key] = {}
             for k in c:
                 try:
-                    self.currentDict[key][k] = float(c[k])
+                    self.newDict[key][k] = float(c[k])
                 except:
-                    self.currentDict[key][k] = c[k]
+                    self.newDict[key][k] = c[k]
